requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
av==13.1.0
azure-core==1.32.0
azure-identity==1.21.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.12.23
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
imageio==2.36.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.27.0
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
medmnist==3.0.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mosspy==1.0.9
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
namex==0.0.8
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
openai==1.70.0
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.2
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pillow==10.4.0
platformdirs==4.3.1
prance==23.6.21.0
prompt_toolkit==3.0.47
propcache==0.3.1
protobuf==5.29.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.11.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==1.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.2
semantic-kernel==1.26.1
setuptools==75.2.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.44.0
stringzilla==3.11.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.1
wheel==0.45.1
wrapt==1.17.0
XlsxWriter==3.2.2
yarl==1.18.3
zipp==3.21.0
